% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_elsoc.R
\name{describe_elsoc}
\alias{describe_elsoc}
\title{describe_elsoc}
\usage{
describe_elsoc(x)
}
\arguments{
\item{x}{Variable name as a string}
}
\description{
Gives a full description of a variable, presenting variable's code, label, question,
module, answers values and label values, and waves and samples in which it was measured
}
