% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_elsoc.R
\name{design_elsoc}
\alias{design_elsoc}
\title{design_elsoc}
\usage{
design_elsoc(
  .data,
  ids = "segmento",
  strata = "estrato",
  weights = "ponderador02",
  nest = TRUE
)
}
\arguments{
\item{ids}{= "segmento"}

\item{strata}{= "estrato"}

\item{weights}{= "ponderador02"}

\item{nest}{= TRUE}
}
\value{

}
\description{
design_elsoc
}
\examples{
elsoc_long_2016_2021 \%>\% design_elsoc()
}
