% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsoc_example.R
\docType{data}
\name{elsoc_example}
\alias{elsoc_example}
\title{Subset of ELSOC 2016-2021 for running examples}
\format{
A data frame with 1525 rows and 8 variables:
\describe{
\item{idencuesta}{casewise id}
\item{ola}{wave of the study}
...
}
}
\usage{
elsoc_example
}
\description{
Subset of ELSOC 2016-2021 for running examples
}
\keyword{datasets}
