% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{freq}
\usage{
freq(.data, var = NULL, by, condition, vartype = "se")
}
\arguments{
\item{var}{= NULL}

\item{vartype}{= 'se'}
}
\value{

}
\description{
Function to create a frecuency table
}
\examples{
elsoc_long_2016_2019 \%>\% freq(by = c(m0_sexo, ola))
}
