% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_nsnr.R
\name{is_nsnr}
\alias{is_nsnr}
\title{is_nsnr}
\usage{
is_nsnr(..., values = c(-777, -888, -999))
}
\arguments{
\item{...}{variables to test}

\item{values}{Values to test if are contained in variables}
}
\description{
Function to test if multiple variables contain the values associated with Doesn't know or Doesn't answer (default = c(-888, -999)).
Useful for testing multiple variables in shorter code
}
\examples{
elsoc_example \%>\% dplyr::filter(have.value(c01, value = c(-777, -888, -999)))

}
