% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_elsoc.R
\name{load_elsoc}
\alias{load_elsoc}
\title{load_elsoc}
\usage{
load_elsoc(data = "long", name = NULL)
}
\arguments{
\item{data}{string specifying format of data to load. Options are:
\itemize{
\item 'long' (default)
\item 'wide'
\item 'territorial-2016'
\item 'territorial-2017'
\item 'territorial-2018'
\item 'territorial-2019'
\item 'elsoc-cit-2016'
}}

\item{name}{specify the name in which the data object is stored}
}
\value{
loads the elsoc dataset into the workspace
}
\description{
Loads the elsoc dataset into the workspace
}
\examples{

load_elsoc()
load_elsoc(name = 'anyname')
load_elsoc(data = 'wide')
load_elsoc(data = 'territorial-2019')

}
