% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookfor.R
\name{lookfor}
\alias{lookfor}
\title{lookfor}
\usage{
lookfor(pattern = ".")
}
\arguments{
\item{pattern}{Character string to be looked for in the ELSOC variable list}
}
\description{
Function to find variable names and label based on a pattern in the ELSOC variable list. Regular expressions for the pattern is supported
}
\examples{

lookfor('c01')
lookfor('cohesion')
}
