% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{stats}
\usage{
stats(
  .data,
  x,
  by = NULL,
  stat = c("mean", "median", "total", "quantile"),
  vartype = c("se", "ci", "var", "cv"),
  quantiles = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{.data}{Data frame or tbl_svy survey object object}

\item{x}{Variable or variable name to calculate estimates}

\item{by}{Vector of variables to group estimates}

\item{stat}{Type of estimate to calculate. One of: 'mean', 'median', 'total' or 'quantile'}

\item{vartype}{Report variability as one or more of: standard error ('se', default), confidence interval ('ci'), variance ('var') or coefficient of variation ('cv'). vartype = NULL for no variability}

\item{na.rmA}{logical value to indicate whether missing values of x should be dropped}
}
\description{
Calculates stats (mean, median, total and quantiles) and its variance from ELSOC considering complex survey design
}
\examples{
elsoc_example \%>\% stat(m0_sexo, by = ola, stat = 'mean')
elsoc_example \%>\% stat(c01, by = c(ola, m0_sexo), stat = 'median')

}
