% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_design_elsoc.R
\name{survey_design_elsoc}
\alias{survey_design_elsoc}
\title{survey_design_elsoc}
\usage{
survey_design_elsoc(
  .data,
  ids = "segmento_disenno",
  strata = "estrato_disenno",
  weights = "ponderador02",
  nest = TRUE,
  ...
)
}
\arguments{
\item{.data}{data frame (ELSOC) in long format.}

\item{ids}{Variable or variable name in .data especifying cluster ids.}

\item{strata}{Variable or variable name in .data especifying strata.}

\item{weights}{Variable or variable name in .data especifying weights. Missing values are droped.}

\item{nest}{if TRUE, relabel cluster ids to enforce nesting within strata}
}
\description{
survey_design_elsoc can be used to create a srvyr's tbl_svy object using design information from ELSOC
}
\examples{
elsoc_example \%>\% survey_design_elsoc()

}
