% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_list.R
\docType{data}
\name{variable_list}
\alias{variable_list}
\title{Variable list of ELSOC databases}
\format{
A data frame with 720 rows and 5 variables
}
\source{
variable_list
}
\usage{
variable_list
}
\description{
Variable list of ELSOC databases
}
\keyword{datasets}
