% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsoc_example.R
\docType{data}
\name{elsoc_example}
\alias{elsoc_example}
\title{ELSOC 2016 2019 reduced to use in examples}
\format{
A data frame with 53940 rows and 10 variables:
}
\source{
elsoc_2016_2019_long
}
\usage{
elsoc_example
}
\description{
ELSOC 2016 2019 reduced to use in examples
}
\keyword{datasets}
