% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{freq}
\usage{
freq(.data, x, by, condition, vartype = "se")
}
\arguments{
\item{.data}{data frame (ELSOC) in long format}

\item{x}{variable, variable name or logical vector to calculate proportions}

\item{by}{vector of variables to group estimates}

\item{vartype}{Report variability as one or more of: standard error ('se', default), confidence interval ('ci'), variance ('var') or coefficient of variation ('cv')}
}
\value{

}
\description{
Calculates proportions and its variance from ELSOC considering complex survey design
}
\examples{
load_elsoc(format = 'long')
elsoc_long_2016_2019 \%>\% freq(by = c(m0_sexo, ola))
}
