% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_elsoc.R
\name{load_elsoc}
\alias{load_elsoc}
\title{load_elsoc}
\usage{
load_elsoc(url = NULL, format = "long")
}
\arguments{
\item{url}{URL specifying directory of database}

\item{format}{string specifying format of data to load}
}
\value{
loads the elsoc dataset into the workspace
}
\description{
Loads the elsoc dataset into the workspace
}
\examples{
load_elsoc()
}
