% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\title{prop}
\usage{
prop(
  .data,
  x,
  by = NULL,
  vartype = c("se", "ci", "var", "cv"),
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{.data}{Data frame or tbl_svy survey object object}

\item{x}{Variable, variable name or logical vector to calculate proportions}

\item{by}{Vector of variables to group estimates}

\item{vartype}{Report variability as one or more of: standard error ('se', default), confidence interval ('ci'), variance ('var') or coefficient of variation ('cv'). vartype = NULL for no variability}

\item{na.rm}{A logical value to indicate whether missing values of x should be dropped}

\item{...}{Other options are pass down to srvyr::survey_mean call}
}
\description{
Calculates proportions and its variance of a variable or condition from ELSOC,
considering complex survey design
}
\examples{
elsoc_example \%>\% prop(m0_sexo, by = ola)
elsoc_example \%>\% prop(c01 \%in\% c(4,5), by = c(ola, m0_sexo))

}
